FIX: compilation warnings

+++ client/gui/src/scenes/World/Player.cpp
@@ -51,7 +51,7 @@ Player::Player(int teamIndex, const raylib::Vector2 &mapSize)
 
 raylib::Color Player::getColor() const
 {
-    if (_teamIdx < 0 || _teamIdx >= COLORS.size())
+    if (_teamIdx < 0 || static_cast<std::size_t>(_teamIdx) >= COLORS.size())
         return raylib::Color::DarkPurple();
     return COLORS[_teamIdx];
 }
